per #86 Add ability to specify alert handlers per alert Wont break existing config. https://github.com/klen/graphite-beacon/issues/86
