+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
 commit = True
-current_version = 0.25.4
+current_version = 0.25.5
 files = graphite_beacon/__init__.py
 tag = True
 tag_name = {new_version}
